Fix filament forms (take 2) (#62)
**Take 2** I posted this as #51 but I had set it up to compare with branch for
#55, rather than `main`, so then when it merged, it went to the branch for #55,
but _after_ that branch had already been merged to `main`. 😩 

Editing posts in the admin site was not straightforward as the resource was not
set up to transfer all the necessary fields, and also the unique constraint on
`slug` would prevent revisions of a post from being saved, due to the way we
maintain prior versions in the same table.

Editing snippets was also not supported as we needed to implement a form.

## Screenshots

**Snippet editing**
https://github.com/user-attachments/assets/f1a52d65-3547-4c2b-a62a-1032fee5ad3d

**Post editing**
https://github.com/user-attachments/assets/f8b0ab58-26c7-4f91-ac5c-fd7e2d580664